[*] Completed base functional api for author entity
Get, save, delete, update, get all — has been done

+++ backend/src/main/java/com/elib/author/AuthorService.java
@@ -2,13 +2,16 @@ package com.elib.author;
 
 import com.elib.author.dto.AuthorDto;
 import com.elib.author.model.Author;
+import com.elib.exception.AuthorNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
 @Service
+@Transactional
 public class AuthorService {
 
     @Autowired
@@ -17,13 +20,13 @@ public class AuthorService {
     public AuthorDto getById(Long id) {
         return authorRepository.findById(id)
                 .map(AuthorMapper.MAPPER::authorToAuthorDto)
-                .orElseThrow(NullPointerException::new);
+                .orElseThrow(AuthorNotFoundException::new);
     }
 
     public AuthorDto getByName(String name) {
         return authorRepository.findByName(name)
                 .map(AuthorMapper.MAPPER::authorToAuthorDto)
-                .orElseThrow(NullPointerException::new);
+                .orElseThrow(AuthorNotFoundException::new);
     }
 
     public List<AuthorDto> getAuthors() {
@@ -36,8 +39,18 @@ public class AuthorService {
         authorRepository.save(author);
     }
 
+    public void update(Long id, String name) {
+        Author author = authorRepository.findById(id)
+                .orElseThrow(AuthorNotFoundException::new);
+        authorRepository.updateName(id, name);
+    }
+
     public void deleteById(Long id) {
-        authorRepository.deleteById(id);
+        if (authorRepository.existsById(id)) {
+            authorRepository.deleteById(id);
+        } else {
+            throw new AuthorNotFoundException();
+        }
     }
 
 }
